requirements.txt
apeye==1.4.1
apeye-core==1.1.5
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
domdf-python-tools==3.9.0
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
natsort==8.4.0
packaging==24.2
platformdirs==4.3.6
pymongo==4.10.1
pypi-json==0.4.0
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
